More Optimization
Added tileTrigger Prefab
----
place object in scene and adjust hit box to fit area you want.

tilesOn and tilesOff are inputs for the tag of the tiles you want to
change. For this to work you must go through the scene and assign all
the tiles with the tag (I did a bunch of them in level3 obviously not
them all as i was just testing it out) you can mess with this and find
the right areas to turn zones on and off to keep the game running really
smooth.
----

slight change to ghost audio to fix it overwriting unity audio
component.

more functionality with movingPlatform script
----
oneShot condition - moves only 1 time to targetB once player enters the
trigger (still needs a targetA to pass error checks, you can set this
anything it doesnt matter)

make sure to adjust the trigger box to fit the area you want the player
on before it moves.
----