Game Update & 1/2 level6
----Important Game update----
Make sure to test play game in web (960x600) as this will be the mode
the game is built in.

Most UI elements are positioned based around this.
----

half of tiles placed for level6, 3 paths to get to midway check point. 2
of them are very difficult and bottom path is much easier.

cleaned up some code, removed some unused things.